Skull Detaching (Read Desc.)
Currently, reattaching is not possible. It will be added in the next commit. Read the note on lines 690-692 for my current idea on how to add it.
Another thing, the way you attach or detach a head currently is by pressing the assigned key (Default is F4) together with the number of the head you want to attach / detach (This is bound to your hotbar slot keybinds)